change the cast when using PRId64 from long long to int64_t

+++ cyphertite/ct_fb.c
@@ -883,7 +883,7 @@ ctfb_readdir_versions(void *arg)
 
 		if (strftime(buf + sz , sizeof(buf) - sz, ".%Y%m%d%H%M%S",
 		    localtime(&mtime)) == 0) {
-			CWARNX("can't format time %" PRId64, (long long)mtime);
+			CWARNX("can't format time %" PRId64, (int64_t)mtime);
 			return (NULL);
 		}
 		CDBG("%s: %s", __func__, buf);
